use the edit buttons on the details page
